remove cWRP from app
